Apply name changes and folder reorganization
